Fix directory mess in computation lib.
