--- computationalLib/include/computationalLib/math_core/math_core.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include "common/commonTypes.h"
-#include "Dipoles.h"
-#include "TestRunner.h"
