+++ computationalLib/include/math_core/math_core.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "common/commonTypes.h"
+#include "Dipoles.h"
+#include "TestRunner.h"
