+++ computationalLib/include/network_shared/AMQPConsumerService.h
