+++ computationalLib/include/network_shared/amqpCommon.h
