+++ computationalLib/include/network_shared/amqpRestService.h
