+++ computationalLib/lib/network_shared/AMQPConsumerService.cpp
