+++ computationalLib/lib/network_shared/AMQPPublisherService.cpp
