+++ computationalLib/lib/network_shared/amqpCommon.cpp
