+++ computationalLib/lib/network_shared/amqpRestService.cpp
