+++ computationalLib/lib/network_shared/networkTypes.cpp
