+++ computationalLib/lib/network_shared/restUtils.cpp
